Color Class added! Tile interaction added!
The color class is added. The user can interact with the tiles and when
all the tile for a word is completed the word will know that and give
the tiles an other color. TODO: Cross recognition between words that
share a tile + when a tile is unclick check if the tile is a tile for
completed word.